refactor: extract track-info into util
