+++ src/components/track-info.js
@@ -1,6 +1,6 @@
 export const createTrackInfo = () => `
-    <div style="display:flex;flex-direction:column;margin-top:0.5rem;">
-        <p id="track-title" style="color:#fff;font-size:14px;"></p>
-        <p id="track-artists" style="font-size:14px;color:#b3b3b3;"></p>
+    <div style="display:flex;flex-direction:column;width:100%;justify-content:space-evenly;margin-top:.5rem">
+        <div class="container" style="width:282px"><div class="track-text" id="track-title"></div></div>
+        <div class="container" style="width:282px"><div class="track-text" id="track-artists"></div></div>
     </div>
 `
