+++ src/popup/index.html
@@ -3,6 +3,7 @@
     <head>
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link href="../styles.css" rel="stylesheet">
         <link href="styles.css" rel="stylesheet">
     </head>
     <body><script type="module" src="index.js"></script></body>
